requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.92.0
httplib2==0.22.0
requests==2.31.0